refactor(core): consolidate tooling architecture
