+++ .changeset/core-tooling-snapshots.md
@@ -1,0 +1,5 @@
+---
+"@crustjs/core": minor
+---
+
+Expose parser-derived flag negation and command-listing policy on the tooling snapshot surface. Move section listing helpers from `@crustjs/core` to `@crustjs/core/tooling` and stop exporting the internal `Simplify` type.
