+++ apps/docs/content/docs/api/index.mdx
@@ -18,8 +18,6 @@ All exports on these pages come from `@crustjs/core`.
 | `defineExtensionId(id)`                                   | Mint a branded Extension identity                              |
 | `defineFlag(name, definition)`                            | Define one named flag, preserving its literal definition       |
 | `defineArg(name, definition)`                             | Define one named positional argument                           |
-| `sectionsFor(sections, consumer)`                         | Select sections visible to the given consumer id               |
-| `visibleSectionsFor(snapshot, consumer)`                  | Collect filtered sections across the visible command tree      |
 | [`CrustError`](/docs/api/errors)                          | Structured framework error with one of four stable codes       |
 
 ## Types
@@ -27,8 +25,9 @@ All exports on these pages come from `@crustjs/core`.
 See [Types](/docs/api/types) for command definitions, Command Snapshots, Command Action contexts, Contexts, Extensions, and error detail types.
 
 <Callout type="warn">
-  `@crustjs/core/tooling` exports the supported `CommandSnapshot` type consumed by
-  [`app.snapshot()`](/docs/api/crust#snapshot). Its remaining helpers — including the
-  presentation-neutral `buildCommandDocumentation()` model used by first-party renderers — are
-  lockstep first-party tooling, not part of the supported application-authoring API.
+  The package root contains invocation-time authoring and execution APIs. `@crustjs/core/tooling`
+  exports the supported `CommandSnapshot` type consumed by
+  [`app.snapshot()`](/docs/api/crust#snapshot), plus build- and render-time helpers such as
+  `isListed()`, `sectionsFor()`, `visibleSectionsFor()`, and the presentation-neutral
+  `buildCommandDocumentation()` model. These helpers move in lockstep with first-party tooling.
 </Callout>
