+++ packages/core/src/command/invocation.ts
@@ -10,13 +10,19 @@ import {
 	type InvocationOutcome,
 } from "../api/extension.ts";
 import { CrustError, type CaughtError } from "../errors.ts";
-import { defineExtensionId, type ExtensionId } from "../identity.ts";
+import type { ExtensionId } from "../identity.ts";
 import { parseArgs, validateParsed } from "../parsing/parser.ts";
 import { applySchemas } from "../parsing/schema.ts";
-import { addFlagSpellingEntries, cloneFlagSpellings } from "../parsing/spellings.ts";
-import { isText } from "../sections.ts";
-import type { CommandSection, FlagDef, FlagsDef, InvocationIO } from "../types.ts";
-import type { CommandDefinition, CrustCommandContext, RunOutcome } from "./crust.ts";
+import type { InvocationIO } from "../types.ts";
+import type { CrustCommandContext, RunOutcome } from "./crust.ts";
+import {
+	applyExtensionCommands,
+	applyExtensionFlags,
+	applyExtensionSections,
+	cloneCommandNode,
+	type MaterializeCommandDefinition,
+	validateAuthoredSections,
+} from "./extensions-install.ts";
 import type { CommandNode } from "./node.ts";
 import { resolveCommand } from "./router.ts";
 import { type CommandSnapshot, snapshotCommand } from "./snapshot.ts";
@@ -32,12 +38,6 @@ interface PreparedInvocation {
 	rootNode: CommandNode;
 	extensions: readonly Extension[];
 }
-
-type MaterializeCommandDefinition = (
-	definition: CommandDefinition,
-	parent: CommandNode,
-	extensionName?: string,
-) => CommandNode;
 
 /**
  * Snapshot subprocess protocol used by first-party build tooling.
@@ -56,102 +56,6 @@ function isAbortError(error: CaughtError): boolean {
 	return error.name === "AbortError";
 }
 
-/** Inject an Extension-owned flag into a node and, when recursive, its descendants. */
-function injectExtensionFlag(
-	node: CommandNode,
-	name: string,
-	def: FlagDef,
-	recursive: boolean,
-): void {
-	// ValidateExtensionFlags owns literal collisions at .extend()/.add(); this
-	// owns dynamic Extensions, where a silent overwrite retypes the owning
-	// command's (or another Extension's) flag at parse time.
-	if (Object.hasOwn(node.effectiveFlags, name)) {
-		throw new CrustError(
-			"DEFINITION",
-			`Extension flag "${name}" collides with a flag already defined on command "${node.meta.name}"`,
-			{ subject: "extension", name, reason: "flag-collision" },
-		);
-	}
-	// The canonical name is checked against the spelling table too: an existing
-	// flag's *alias* equal to the incoming canonical would otherwise be silently
-	// stolen (effectiveFlags only has canonical keys).
-	for (const spelling of [name, def.short, ...(def.aliases ?? [])]) {
-		const existing = spelling === undefined ? undefined : node.flagSpellings.get(spelling);
-		if (existing !== undefined && existing.canonicalName !== name) {
-			throw new CrustError(
-				"DEFINITION",
-				`Extension flag spelling "${spelling}" collides with existing flag "${existing.canonicalName}" on command "${node.meta.name}"`,
-				{ subject: "extension", name, reason: "flag-collision" },
-			);
-		}
-	}
-	node.effectiveFlags[name] = def;
-	addFlagSpellingEntries(node.flagSpellings, name, def);
-	if (!recursive) return;
-	for (const sub of Object.values(node.subCommands)) {
-		injectExtensionFlag(sub, name, def, true);
-	}
-}
-
-/** Attach one Extension's owned root commands to a cloned tree. */
-function applyExtensionCommands(
-	root: CommandNode,
-	extension: Extension,
-	materializeCommandDefinition: MaterializeCommandDefinition,
-): void {
-	for (const definition of extension.commands ?? []) {
-		const node = materializeCommandDefinition(definition, root, extension.id);
-		root.subCommands[definition.name] = node;
-	}
-}
-
-/** Inject one Extension's owned flags across a cloned tree. */
-function applyExtensionFlags(root: CommandNode, extension: Extension): void {
-	for (const [name, defWithScope] of Object.entries(extension.flags ?? {})) {
-		const { recursive = true, ...def } = defWithScope;
-		injectExtensionFlag(root, name, def, recursive);
-	}
-}
-
-function cloneFlags(flags: FlagsDef): FlagsDef {
-	const out: FlagsDef = {};
-	for (const [key, def] of Object.entries(flags)) {
-		out[key] = {
-			...def,
-			aliases: def.aliases ? [...def.aliases] : undefined,
-		};
-	}
-	return out;
-}
-
-/** Deep-clone a command subtree without mutating the builder graph. */
-export function cloneCommandNode(node: CommandNode): CommandNode {
-	const subCommands: Record<string, CommandNode> = {};
-	for (const [name, sub] of Object.entries(node.subCommands)) {
-		subCommands[name] = cloneCommandNode(sub);
-	}
-
-	const effectiveFlags = cloneFlags(node.effectiveFlags);
-	// Spread first, then override every structural field with a decoupled copy.
-	return {
-		...node,
-		// Section objects/arrays are never mutated in place (prepare replaces
-		// them wholesale), so sharing them here is safe.
-		meta: { ...node.meta },
-		localFlags: cloneFlags(node.localFlags),
-		ownedFlags: cloneFlags(node.ownedFlags),
-		effectiveFlags,
-		flagSpellings: cloneFlagSpellings(node.flagSpellings, effectiveFlags),
-		args: node.args ? node.args.map((def) => ({ ...def })) : undefined,
-		subCommands,
-		contexts: [...node.contexts],
-		contextExtensionIds: [...node.contextExtensionIds],
-		extensions: [...node.extensions],
-		run: node.run,
-	};
-}
-
 function freezeTree(node: CommandNode): void {
 	Object.freeze(node);
 	Object.freeze(node.localFlags);
@@ -168,126 +72,8 @@ function freezeTree(node: CommandNode): void {
 	Object.freeze(node.subCommands);
 }
 
-/** Who authored the sections being validated; error labels derive from this. */
-type SectionOwner = { subject: "command" | "extension"; name: string };
-
-function invalidSections({ subject, name }: SectionOwner): CrustError {
-	const label = subject === "command" ? "Command" : "Extension";
-	return new CrustError(
-		"DEFINITION",
-		`${label} "${name}" contains invalid documentation sections`,
-		{ subject, name, reason: "invalid-sections" },
-	);
-}
-
-function hasId<T>(value: T): value is T & { readonly id?: unknown } {
-	return (
-		((typeof value === "object" && value !== null) || typeof value === "function") && "id" in value
-	);
-}
-
-function isString<T>(value: T): value is T & string {
-	return typeof value === "string";
-}
-
 function isSymbol<T>(value: T): value is T & symbol {
 	return typeof value === "symbol";
-}
-
-function parseExtensionId(consumer: unknown, owner: SectionOwner): ExtensionId {
-	const id = isString(consumer) ? consumer : hasId(consumer) ? consumer.id : undefined;
-	if (!isText(id) || id !== id.trim()) throw invalidSections(owner);
-	return defineExtensionId(id);
-}
-
-function validateSectionAudienceIds(ids: unknown, owner: SectionOwner): readonly ExtensionId[] {
-	if (!Array.isArray(ids) || ids.length === 0) throw invalidSections(owner);
-	return Object.freeze(ids.map((consumer) => parseExtensionId(consumer, owner)));
-}
-
-function validateSection(section: unknown, owner: SectionOwner): CommandSection {
-	// SAFETY: optional-field probe of an unvalidated section; every field is checked below.
-	const { title, body, only, except } = (section ?? {}) as {
-		title?: unknown;
-		body?: unknown;
-		only?: unknown;
-		except?: unknown;
-	};
-	if (!isText(title) || /[\r\n]/.test(title) || !isText(body)) {
-		throw invalidSections(owner);
-	}
-	// The SectionAudience union owns literals; this runtime branch owns the
-	// dynamic path (Extension `sections` callbacks, config-built objects), where
-	// both fields would otherwise freeze and `sectionsFor()` would silently
-	// ignore `except`.
-	if (only !== undefined && except !== undefined) {
-		throw invalidSections(owner);
-	}
-	if (only !== undefined) {
-		return Object.freeze({ title, body, only: validateSectionAudienceIds(only, owner) });
-	}
-	if (except !== undefined) {
-		return Object.freeze({ title, body, except: validateSectionAudienceIds(except, owner) });
-	}
-	return Object.freeze({ title, body });
-}
-
-function validateAuthoredSections(node: CommandNode): void {
-	const sections = node.meta.sections;
-	if (sections !== undefined) {
-		const owner: SectionOwner = { subject: "command", name: node.meta.name };
-		if (!Array.isArray(sections)) throw invalidSections(owner);
-		node.meta.sections = sections.map((section) => validateSection(section, owner));
-	}
-	for (const sub of Object.values(node.subCommands)) validateAuthoredSections(sub);
-}
-
-function contributionTarget(
-	root: CommandNode,
-	command: readonly string[],
-	extension: Extension,
-): CommandNode {
-	let target = root;
-	for (const segment of command) {
-		// hasOwn: plain-object lookup would resolve inherited keys like "constructor"
-		const next = Object.hasOwn(target.subCommands, segment)
-			? target.subCommands[segment]
-			: undefined;
-		if (!next) {
-			throw new CrustError(
-				"DEFINITION",
-				`Extension "${extension.id}" section target "${command.join(" ")}" is not a canonical command path`,
-				{
-					subject: "extension",
-					name: extension.id,
-					reason: "invalid-section-path",
-				},
-			);
-		}
-		target = next;
-	}
-	return target;
-}
-
-function applyExtensionSections(
-	root: CommandNode,
-	extension: Extension,
-	snapshot: CommandSnapshot,
-): void {
-	if (!extension.sections) return;
-	const owner: SectionOwner = { subject: "extension", name: extension.id };
-	const contributions = extension.sections(snapshot);
-	if (!Array.isArray(contributions)) throw invalidSections(owner);
-	for (const contribution of contributions) {
-		// validateSection rejects null/non-object contributions, so reading
-		// `.command` afterwards is safe.
-		const section = validateSection(contribution, owner);
-		if (!Array.isArray(contribution.command) || !contribution.command.every(isString)) {
-			throw invalidSections(owner);
-		}
-		const target = contributionTarget(root, contribution.command, extension);
-		target.meta.sections = [...(target.meta.sections ?? []), section];
-	}
 }
 
 const preparedInvocations = new WeakMap<CommandNode, PreparedInvocation>();
